Add a reducer that adds book fixing bugs and rendering the book store
